+++ test/utils/string_exensions_test.dart
@@ -14,6 +14,8 @@ void tests() {
       expect('ARA1'.toSnakeCase(), 'ara_1');
       expect('start15min'.toSnakeCase(), 'start_15min');
       expect('sch_bid_y1axis_data'.toSnakeCase(), 'sch_bid_y1_axis_data');
+      expect('so2_mass'.toSnakeCase(), 'so2_mass');
+      expect('so2Mass'.toSnakeCase(), 'so2_mass');
     });
 
   });